backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv==1.0.1

# LangGraph & LangChain
langgraph==0.0.50
langchain==0.1.9
langchain-community==0.0.24

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Document Generation
python-docx==1.1.0

# Logging
loguru==0.7.2

# Utilities
jinja2==3.1.3
pyyaml==6.0.1
